requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.3.0
dj-rest-auth>=5.0.0
django-allauth>=0.57.0
requests>=2.31.0
gunicorn>=21.2.0
google-auth>=2.23.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
